[Norm] Finished normalize data MercadoLibre and Amazon - 2020/04/06
